+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Web SDK Change Log
 
+## 3.0.3
+
+* Adds a `setImmediate` function that defers execution without using `setTimeout` but uses a `setTimeout` fallback.
+* Fixes the `move` event for `layer.Query`.
+
 ## 3.0.2
 
 * Fixes error in `npm publish`